+++ crates/utils/src/settings/mod.rs
@@ -6,6 +6,6 @@ pub enum SettingsProviderError {
     Internal(#[source] Box<dyn std::error::Error + Send + Sync + 'static>),
 }
 
-pub trait Settings {
+pub trait Settings: Send {
     fn get_settings_or_panic(&self, name: &'static str) -> String;
 }